+++ update-provider-configuration.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-#  v2.7.0
+#  v2.9.0
 # reduced sleep during downtime to 5 seconds
 
 set -euo pipefail
@@ -39,3 +39,28 @@ kubectl -n akash-services scale statefulsets akash-provider --replicas=1
 sleep 15
 echo "step to verify it's been started."
 kubectl -n akash-services get statefulsets && kubectl -n akash-services get pods -l app=akash-provide
+
+# This turned out not to be useful, since code in other scripts wasn't viable.
+#echo "saving new rpc node to tmp/akash-provider-rpc.conf"
+# RAW=$(kubectl exec -n akash-services akash-provider-0 -c provider -- \
+#            printenv AKASH_NODE_1_PORT_26657_TCP 2>/dev/null)
+
+ #         if [[ -z "$RAW" ]]; then
+ #           echo "No AKASH_NODE_1_PORT_26657_TCP found"
+ #           exit 1
+ #         fi
+
+ #         HOST=$(echo "$RAW" | sed -E 's#^tcp://([^:]+):([0-9]+)$#\1#')
+ #         PORT=$(echo "$RAW" | sed -E 's#^tcp://([^:]+):([0-9]+)$#\2#')
+
+ #         if [[ "$PORT" == "26657" ]]; then
+ #           SCHEME="http"
+ #         elif [[ "$PORT" == "440" ]]; then
+ #           SCHEME="https"
+ #         else
+ #           SCHEME="https"
+ #         fi
+
+  #        rpc_node_url="${SCHEME}://${HOST}:${PORT}"
+  #        echo "$rpc_node_url" > /tmp/akash-provider-rpc.conf
+