+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
 Most notable changes to this project, should be documented in this file.
+## v2.9.0	
+		Lots of changes, mainly ticker.sh, added new check script in scripts folder.
+		update-provider-configuration.sh also changed, added so it will update the current-provider-rpc.tmp file.
+
+## v2.8.6
+		Changed logging by ticker.sh during early break in sleep.
+		Now shows reasons for provider pod restarts.
+		Error code check replaced with kubectl event logs for provider pod.
+
+## v2.8.5
+		Bug fix, making the 2.8.0 - Reduced Provider Pod restart detection feature actually immediate.
+		minor ticker.sh code change
+
 ## v2.8.4 	added --version 11.6.4 to helm upgrade in update-provider-yaml.sh
 
 ## v2.8.2	bug fix in akash-provider-console